Fix clang-tidy and clang-format linting errors
- Add explicit static_cast<int> for narrowing conversions from
  sqlite3_int64 in db_manager.cpp (x2), int64_t in api_routes.cpp (x2),
  config_parser.cpp, and google_oauth.cpp
- Fix bugprone-misplaced-widening-cast in token_encryption.cpp:
  cast KEY_LEN to size_t before multiplication, and IV_LEN before addition
- Fix bugprone-implicit-widening-of-multiplication-result: cast loop index
  i to size_t before multiplying when indexing keyHex
- Fix narrowing conversion from unsigned char to char in two
  BASE64_CHARS.find() calls in token_encryption.cpp
- Fix middleware.h after_handle signature to use context& instead of
  const context& to match Crow's arity-3 type trait
- Apply clang-format to all touched files

+++ src/api_routes.cpp
@@ -80,7 +80,7 @@ void setupRoutes(crow::App<RequestTimerMiddleware> &app, std::shared_ptr<DBManag
                 for (const auto &ingJson : body["ingredients"]) {
                     std::string ingName = ingJson["name"].s();
                     double amount = ingJson["amount"].d();
-                    int unit = ingJson["unit"].i();
+                    int unit = static_cast<int>(ingJson["unit"].i());
                     // Preparation is optional in the JSON
                     std::string prep = "None";
                     if (ingJson.has("preparation")) {
@@ -132,7 +132,7 @@ void setupRoutes(crow::App<RequestTimerMiddleware> &app, std::shared_ptr<DBManag
                     for (const auto &ingJson : body["ingredients"]) {
                         std::string ingName = ingJson["name"].s();
                         double amount = ingJson["amount"].d();
-                        int unit = ingJson["unit"].i();
+                        int unit = static_cast<int>(ingJson["unit"].i());
                         std::string prep = "None";
                         if (ingJson.has("preparation")) {
                             prep = ingJson["preparation"].s();
